ml-skillgram-recsys/requirements.txt
numpy>=1.26
pandas>=2.1
scipy>=1.11
scikit-learn>=1.4
implicit>=0.7.2
sentence-transformers>=2.7
faiss-cpu>=1.8
pyarrow>=15
requests>=2.31
tqdm>=4.66
python-dotenv>=1.0
